backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pandas>=2.0.0
scikit-learn>=1.4.0
umap-learn>=0.5.6
numpy>=1.26.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
